Extract monitoring tests to separate file for 400-line compliance
Move the test module from monitoring.rs (419 lines) to a dedicated
monitoring_tests.rs file (227 lines), reducing the main module to 197 lines.

Changes:
- Create monitoring_tests.rs with all test functions
- Expose HealthCheckOutcome, ProcessMonitorContext, and helper functions
  with pub(crate) visibility for test access
- Add #[cfg(test)] mod monitoring_tests declaration to mod.rs
